完善各种试验代码

+++ 05_位置式PID定速控制-积分限幅/MiddleWare/MOTOR/motor.h
@@ -9,7 +9,7 @@
 
 
 
-extern void Motor_SetPWM(int8_t PWM);
+extern void Motor_SetPWM(int16_t PWM);
 extern int16_t Encoder_Get(void);
 #endif
